Fix VS crash when image files during slideshow removed, fix naming of SlideShowImageProvider
